Add controllers

+++ src/main/java/com/zaurtregulov/spring/data/repository/MaintenanceScheduleRepositoryImpl.java
@@ -3,6 +3,6 @@ package com.zaurtregulov.spring.data.repository;
 import com.zaurtregulov.spring.data.entity.MaintenanceSchedule;
 import com.zaurtregulov.spring.domain.repository.MaintenanceScheduleRepository;
 
-public class MaintenanceScheduleRepositoryImpl extends BaseRepository<MaintenanceSchedule, Integer> implements MaintenanceScheduleRepository {
+public class MaintenanceScheduleRepositoryImpl extends BaseRepository<MaintenanceSchedule, Long> implements MaintenanceScheduleRepository {
 
 }
